api/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
# Azure Search and Prompt Flow dependencies
azure-search-documents
azure-core
promptflow
promptflow-tools
keyrings.alt
openai
# Production server
gunicorn
# File upload support
python-multipart
# CV processing
PyPDF2==3.0.1
python-docx==1.1.0
# User management and authentication
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
